mac : update cmake_build.sh : use a new CMakeLists.txt file in order to make install into package, need to be edit to make it more editable (change package name easily for example), and locale problem still there.

+++ trunk/mac/cmake_build.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 
+#################################################
+#  Script for compile with cmake under MacOS X  #
+#           For     W O R M U X                 #
+#    done by Plorf, Auria and lynxlynxlynx      #
+#################################################
+
+
 #
 # Set environment for compile
 #
@@ -7,14 +14,82 @@
 MAC=`pwd`/
 ROOT=${MAC}../
 SRC=${ROOT}src/
-cp libSDLmain_UB.a ${ROOT}
-cp libintl.a ${ROOT}
-mkdir ${MAC}tmpbuild
+
+
+#
+# Set files for CMake and compilation
+#
+
+# Copy the library libSDLmain_UB.a
+if [ -e libSDLmain_UB.a ]
+then
+    cp libSDLmain_UB.a ${ROOT}
+else
+    echo "*** ${MAC}libSDLmain_UB.a Not Found ***"
+    exit
+fi
+
+# Copy the librarie libintl.a
+if [ -e libintl.a ]
+then
+    cp libintl.a ${ROOT}
+else
+    echo "*** ${MAC}libintl.a Not Found ***"
+    exit
+fi
+
+# Set the good release of CMakeLists
+if [ -e CMakeLists.mac.txt ]
+then
+    cd ${ROOT}
+    mv ${ROOT}CMakeLists.txt ${ROOT}tmp.$$
+    cp ${MAC}CMakeLists.mac.txt ${ROOT}CMakeLists.txt
+else
+    echo "*** ${MAC}CMakeLists.mac.txt Not Found ***"
+    exit
+fi
+
+
+#
+# Clean temps files, and create Folders
+#
+
 TMP=${MAC}tmpbuild/
-mkdir ${MAC}tmpdata
-TMPDATA=${MAC}tmpdata/
+if [ -e ${TMP} ]
+then
+    echo "*****************"
+    echo "  Clean tmpbuild"
+#    rm -rf ${TMP}*
+    echo "*****************"
+else
+    mkdir ${TMP}
+fi
+
+APP=${MAC}Wormux.app/
+if [ -e ${APP} ]
+then
+    echo "******************"
+    echo "Clean Package .app"
+    rm -rf ${APP}/*
+    echo "******************"
+fi
+echo "Create Wormux.app file"
+mkdir -p ${APP}
+mkdir -p ${APP}Contents/MacOS/
+mkdir -p ${APP}Contents/Frameworks/
+RES=${APP}Contents/Resources/
+mkdir -p ${RES}data/
+mkdir -p ${RES}locale/
+
 
 ARCHIVE=${MAC}Wormux-0.8svn.tar.bz2
+if [ -e ${ARCHIVE} ]
+then
+    echo "******************"
+    echo "Remove old Archive" 
+    rm -f ${ARCHIVE}
+    echo "******************"
+fi
 
 export CMAKE_INSTALL_PREFIX=./wormux-files
 
@@ -23,48 +98,26 @@ export CMAKE_INSTALL_PREFIX=./wormux-files
 #
 
 cd ${TMP}
-
-awk '/^SET\(WORMUX_PATCH/ { sub(/^/,"#") } { print }' ${ROOT}CMakeLists.txt > tmp.$$.$$
-cp ${ROOT}CMakeLists.txt tmp.$$.$$.2
-mv tmp.$$.$$ ${ROOT}CMakeLists.txt
-#cmake ../.. -DPREFIX:PATCH=${TMPDATA}
+#awk '/^SET\(WORMUX_PATCH/ { sub(/^/,"#") } { print }' ${ROOT}CMakeLists.txt > tmp.$$.$$
+#cp ${ROOT}CMakeLists.txt tmp.$$.$$.2
+#mv tmp.$$.$$ ${ROOT}CMakeLists.txt
 cmake ../..
-mv tmp.$$.$$.2 ${ROOT}CMakeLists.txt
+#mv tmp.$$.$$.2 ${ROOT}CMakeLists.txt
 
 make
 make install
 
 #
-# Clean environment
-#
-
-rm ${ROOT}/libSDLmain_UB.a ${ROOT}/libintl.a 
-
-
-#
 # Generate .app File
 #
 
-echo "Create Wormux.app file"
-
-APP=${MAC}Wormux.app/
-rm -rf ${APP} ${ARCHIVE}
-
-mkdir -p ${APP}
-mkdir -p ${APP}Contents/Resources/
-mkdir -p ${APP}Contents/MacOS/
-mkdir -p ${APP}Contents/Frameworks/
-RES=${APP}Contents/Resources/
-mkdir -p ${RES}data/
-mkdir -p ${RES}locale/
-
-echo "Copy data into .app file"
+#echo "Copy data into .app file"
 
 # Copy data files into .app
-echo "Copy /data"
-#cp -R ${TMPDATA}share/wormux/* ${RES}data/
-INSTALL=/usr/local/share/
-cp -R ${INSTALL}wormux/* ${RES}data
+#echo "Copy /data"
+#cp -R ${TMPDATA}wdata/* ${RES}data/
+#INSTALL=/usr/local/share/
+#cp -R ${INSTALL}wormux/* ${RES}data
 
 # Copy locale files into .app
 #echo "Error : wormux in english, so i think it's not working... search where is it"
@@ -77,26 +130,44 @@ cp -R ${INSTALL}wormux/* ${RES}data
 #make -w -C ${ROOT}/po -e install-data-local-yes localedir=${RES}/locale
 
 TMPLOCALE=;
-for TMPLOCALE in /usr/share/locale/??;
-do [[ -e $lang/LC_MESSAGES/wormux.mo ]] || continue;
-    mkdir -p $target/$lang/LC_MESSAGES;
-    cp $lang/LC_MESSAGES/wormux.mo $target/$lang/LC_MESSAGES/wormux.mo;
-done
+#for TMPLOCALE in /usr/share/locale/??;
+#do [[ -e $lang/LC_MESSAGES/wormux.mo ]] || continue;
+#    mkdir -p $target/$lang/LC_MESSAGES;
+#    cp $lang/LC_MESSAGES/wormux.mo $target/$lang/LC_MESSAGES/wormux.mo;
+#    mkdir -p ${RES}locale/$lang/LC_MESSAGES/wormux.mo
+#    cp $lang/LC_MESSAGES/wormux.mo ${RES}locale/$lang/LC_MESSAGES/wormux.mo
+#done 
+
 
 # Add icon and info.plist and PkgInfo
 cp ${MAC}Info.plist.in ${APP}Contents/info.plist
 cp ${MAC}PkgInfo.in ${APP}Contents/
 cp ${ROOT}data/wormux_128x128.icns ${RES}Wormux.icns
 
+# Do a simple test for check if data is well copied
+if [ -e ${RES}/data/wormux_default_config.xml ]
+then
+    echo "Default_config OK"
+else
+    echo "*** ERROR : No Default_config ! : Probably no datas copied ***"
+    exit 
+fi
 
-echo "Add bin"
-cp ${TMP}src/wormux ${APP}Contents/MacOS/wormux
+
+#echo "Add bin"
+#cp ${TMP}src/wormux ${APP}Contents/MacOS/wormux
 #cp ${TMPDATA}bin/wormux ${APP}Contents/MacOs/wormux
+
+
+#
+# Copy frameworks into package
+#
 
 echo "Copy all frameworks"
 cd ${MAC};
-MIRROR=http://plorf.homeip.net/wormux/lib/
 
+# If frameworks are not available, they'll be download from this mirror
+MIRROR=http://plorf.homeip.net/wormux/lib/
 if [ ! -e "${MAC}frameworks.tar.bz2" ]
 then 
     echo "Frameworks will be downloaded from ${MIRROR} (3MB)";
@@ -105,16 +176,44 @@ fi
     tar xfj ${MAC}frameworks.tar.bz2 -C ${APP}Contents/Frameworks;
     echo "Frameworks copy done"
 
+
 #
 # Make .dmg and .tar.bz2 file
+#
+
+#
+# TODO THERE MAKE .DMG
 #
 
     echo "Make archive ${ARCHIVE}"
     #tar cfj ${ARCHIVE} ${APP}
     #echo "Archive ${ARCHIVE} done"
-
     
-echo "Remove temps files"
+
+#
+# Clean environment
+#
+
+# Remove copy of libSDLmain_UB.a
+if [ -e ${ROOT}libSDLmain_UB.a ]
+then
+    rm ${ROOT}libSDLmain_UB.a
+fi
+
+# Remove copy of libintl.a
+if [ -e ${ROOT}libintl.a ]
+then
+    rm ${ROOT}libintl.a
+fi
+
+# Remove and reset default CMakeLists.txt
+if [ -e ${ROOT}tmp.$$ ]
+then
+    mv ${ROOT}tmp.$$ ${ROOT}CMakeLists.txt
+fi
+
+# Remove tmp files
 #rm -rf ${MAC}tmpbuild
+
 echo "Build done"
